pubsub-homework/requirements.txt
google-cloud-pubsub
pyyaml